previously coded files
These are all the files I wrote when I was coding in my eclipse environment. They comprise a rudimentary Minesweeper game in the form of a Java applet built with Swing. The game can be played with any board size and any (sane) number of mines, and includes functionality such as expanding areas of the board filled with zeros and starting a new game.